refactor(my_steach): Phase 0 - Remove dead code and fix memory leak
Clean up codebase as part of comprehensive plugin review and fix plan:

Changes:
- Remove dead function declarations (launch_panorama_kernel_textured, cleanup_texture_resources)
- Fix type narrowing in output index calculation (cast before arithmetic)
- Fix static CUDA event memory leak (move to struct member with proper lifecycle)
- Keep async color correction struct members (needed for Phase 1)

Impact:
- Removes 21 lines of dead code from header
- Eliminates CUDA event memory leak
- Fixes potential integer overflow in kernel
- Reduces compiler warnings (-Wall -Wextra -Wconversion)

Files modified:
- src/cuda_stitch_kernel.h: Remove unused declarations
- src/cuda_stitch_kernel.cu: Fix type narrowing (line 483-485)
- src/gstnvdsstitch.h: Add frame_complete_event member
- src/gstnvdsstitch.cpp: Proper event lifecycle management

Part of: Plugin review remediation plan
Next: Phase 1 (Async color correction implementation)

+++ my_steach/src/cuda_stitch_kernel.cu
@@ -480,8 +480,9 @@ __global__ void panorama_lut_kernel(
     }
     
     // Записываем результат (с переворотом как у вас было)
-    size_t out_idx = (size_t)(output_height - 1 - y) * (size_t)output_pitch + 
-                 (size_t)(output_width - 1 - x) * 4;
+    // Cast to size_t before arithmetic to prevent integer overflow
+    size_t out_idx = ((size_t)output_height - 1 - (size_t)y) * output_pitch +
+                     ((size_t)output_width - 1 - (size_t)x) * 4;
     
     *((uchar4*)(output + out_idx)) = make_uchar4(
         __float2uint_rn(result.x),
